+++ src/DocumentView.hpp
@@ -7,7 +7,7 @@
 #include "GraphicsView.hpp"
 #include "JumpMarker.hpp"
 #include "Model.hpp"
-#include "VerticalScrollBar.hpp"
+#include "ScrollBar.hpp"
 
 #ifdef HAS_SYNCTEX
 extern "C"
@@ -320,8 +320,8 @@ private:
     float m_spacing{10.0f}, m_page_stride{0.0f}, m_page_x_offset{0.0f};
     double m_target_zoom{1.0}, m_current_zoom{1.0}, m_rotation{0.0};
     bool m_auto_resize{false}, m_auto_reload{false};
-    QScrollBar *m_hscroll{nullptr};
-    VerticalScrollBar *m_vscroll{nullptr};
+    ScrollBar *m_hscroll{nullptr};
+    ScrollBar *m_vscroll{nullptr};
     QHash<int, GraphicsPixmapItem *> m_page_items_hash;
     QHash<int, std::vector<BrowseLinkItem *>> m_page_links_hash;
     QHash<int, std::vector<Annotation *>> m_page_annotations_hash;